refactor(ui): extract api/internal.ts — payloadError, fetchJson, readJsonPayload, buildProjectParams
Second slice of the api.ts decomposition. Move the private fetch
helpers into api/internal.ts so the per-domain modules (memories,
coordinator-admin, sync, etc.) can reuse them without pulling through
the barrel.
